ace/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "learn-ai-ace"
version = "0.1.0"
description = "ACE — Autonomous Curriculum Engine for Learn AI With Grey8"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "jsonschema>=4.21",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
ace = "ace.cli:cli"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
